Removed documentation for secure installs.
We will continue to support secure installs, for the time being, but
not mention it in the documentation. Secure install support, while
provided by Ruby Gems, is not widely used or supported. Due to this,
it's not worth supporting either as it leads to issue reports not
always related to this gem. It would be nice to see this more widely
supported at some point in the future, though.

[ci skip]