Fixed prepending of new lines when table of contents was empty
When attempting to build the Table of Contents for files which had no
headings at all, a new line was being prepended to these files. This
would cause these files to grow in size with a massive number of empty
lines. This fix ensures the contents of these files remain unchanged in
these situations.