Removed build namespace from default configuration
Necessary to bring this more in line with what is happening in the
Pragmater project where there is potential to support both the
insertion and removal of table of contents from documents. For now,
only the ability to insert/update table of contents is supported.

This also has the added benefit of making the configuration easier to
use without the need for nested keys.

The CLI `--build` command has been switched in favor of using
`--insert`. Again, to be more in line with Pragmater usage and provide
support for adding a potentiall `--remove` command in the future.